+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Optimize products mass deletion
 - Improve get product REST API
 - Improve entity history context display for entities updated during import jobs
+- Dispatch event post handler on the mass action dispatcher
 
 ## Bug fixes
 - Fixed verbose option always used